Classes previously located in the MonteCarlo file were now extracted in their own separate files.
