+++ src/eu/veldsoft/vitosha/poker/odds/.gitignore
@@ -1,0 +1,3 @@
+/AboutActivity.class
+/SplashScreen.class
+/VitoshaPokerOddsActivity.class
