+++ src/eu/veldsoft/vitosha/poker/odds/model/.gitignore
@@ -1,0 +1,9 @@
+/Board.class
+/Card.class
+/Constants.class
+/Deck.class
+/Game.class
+/HandStrength.class
+/MonteCarlo.class
+/Player.class
+/Pot.class
