+++ src/eu/veldsoft/vitosha/poker/odds/model/Card.java
@@ -1,0 +1,15 @@
+package eu.veldsoft.vitosha.poker.odds.model;
+
+/**
+ * 
+ * @author Todor Balabanov
+ * 
+ * @email tdb@tbsoft.eu
+ * 
+ * @date 09 Aug 2012
+ */
+class Card {
+	int suit;
+	int kind;
+	boolean known;
+}
