+++ src/eu/veldsoft/vitosha/poker/odds/model/MonteCarlo.java
@@ -34,321 +34,6 @@ import java.util.Random;
  * 
  * @date 09 Aug 2012
  */
-class Constants {
-	static final int HUMAN_PLAYER_INDEX = 0;
-
-	static final long MIN_NUMBER_OF_LOOPS = 1L;
-	static final long MAX_NUMBER_OF_LOOPS = 100000L;
-
-	static final int MIN_NUMBER_OF_PLAYERS = 2;
-	static final int MAX_NUMBER_OF_PLAYERS = 10;
-
-	static final int NUMBER_OF_CARDS = 52;
-	static final int NUMBER_OF_EVALUATION_HAND_CARDS = 5;
-	static final int NUMBER_OF_POSSIBLE_HANDS_FOR_EVALUATION = 21;
-	static final int NUMBER_OF_HAND_CARDS = 2;
-	static final int NUMBER_OF_FLOP_CARDS = 3;
-	static final int NUMBER_OF_KNOWN_CARDS = 7;
-
-	static final int PLAYER_TYPE_REGULAR = 0x1;
-	static final int PLAYER_TYPE_DEALER = 0x2;
-	static final int PLAYER_TYPE_SMALL_BLIND = 0x4;
-	static final int PLAYER_TYPE_BIG_BLIND = 0x8;
-
-	static final long MIN_PLAYER_MONEY = 100L;
-	static final long MAX_PLAYER_MONEY = 100000L;
-
-	static final int CARD_KIND_WEAK_ACE = 0x1;
-	static final int CARD_KIND_TWO = 0x2;
-	static final int CARD_KIND_THREE = 0x3;
-	static final int CARD_KIND_FOUR = 0x4;
-	static final int CARD_KIND_FIVE = 0x5;
-	static final int CARD_KIND_SIX = 0x6;
-	static final int CARD_KIND_SEVEN = 0x7;
-	static final int CARD_KIND_EIGHT = 0x8;
-	static final int CARD_KIND_NINE = 0x9;
-	static final int CARD_KIND_TEN = 0xA;
-	static final int CARD_KIND_JACK = 0xB;
-	static final int CARD_KIND_QUEEN = 0xC;
-	static final int CARD_KIND_KING = 0xD;
-	static final int CARD_KIND_ACE = 0xE;
-
-	static final int CARD_SUIT_CLUBS = 0x10;
-	static final int CARD_SUIT_DIAMONDS = 0x20;
-	static final int CARD_SUIT_HEARTS = 0x30;
-	static final int CARD_SUIT_SPADES = 0x40;
-}
-
-/**
- * 
- * @author Todor Balabanov
- * 
- * @email tdb@tbsoft.eu
- * 
- * @date 09 Aug 2012
- */
-class Card {
-	int suit;
-	int kind;
-	boolean known;
-}
-
-/**
- * 
- * @author Todor Balabanov
- * 
- * @email tdb@tbsoft.eu
- * 
- * @date 09 Aug 2012
- */
-class Deck {
-	Card[] cards = new Card[Constants.NUMBER_OF_CARDS];
-
-	/**
-	 * 
-	 * @author Todor Balabanov
-	 * 
-	 * @email tdb@tbsoft.eu
-	 * 
-	 * @date 09 Aug 2012
-	 */
-	public Deck() {
-		for (int i = 0; i < cards.length; i++) {
-			cards[i] = new Card();
-		}
-	}
-}
-
-/**
- * 
- * @author Todor Balabanov
- * 
- * @email tdb@tbsoft.eu
- * 
- * @date 09 Aug 2012
- */
-class Player {
-	Card[] hand = new Card[Constants.NUMBER_OF_HAND_CARDS];
-	long money;
-	int type;
-	long handScore;
-	long winsCounter;
-	long tiesCounter;
-
-	/**
-	 * 
-	 * @author Todor Balabanov
-	 * 
-	 * @email tdb@tbsoft.eu
-	 * 
-	 * @date 09 Aug 2012
-	 */
-	public Player() {
-		for (int i = 0; i < hand.length; i++) {
-			hand[i] = new Card();
-		}
-	}
-}
-
-/**
- * 
- * @author Todor Balabanov
- * 
- * @email tdb@tbsoft.eu
- * 
- * @date 09 Aug 2012
- */
-class Pot {
-	long[] bets = new long[Constants.MAX_NUMBER_OF_PLAYERS];
-	int numberOfSplits;
-}
-
-/**
- * 
- * @author Todor Balabanov
- * 
- * @email tdb@tbsoft.eu
- * 
- * @date 09 Aug 2012
- */
-class Board {
-	Card[] flop = new Card[Constants.NUMBER_OF_FLOP_CARDS];
-	Card turn = new Card();
-	Card river = new Card();
-
-	/**
-	 * 
-	 * @author Todor Balabanov
-	 * 
-	 * @email tdb@tbsoft.eu
-	 * 
-	 * @date 09 Aug 2012
-	 */
-	public Board() {
-		for (int i = 0; i < flop.length; i++) {
-			flop[i] = new Card();
-		}
-	}
-}
-
-/**
- * 
- * @author Todor Balabanov
- * 
- * @email tdb@tbsoft.eu
- * 
- * @date 09 Aug 2012
- */
-class Game {
-	Deck deck = new Deck();
-	Board board = new Board();
-	Player[] players = new Player[Constants.MAX_NUMBER_OF_PLAYERS];
-	int numberOfPlayers;
-	Pot pot = new Pot();
-
-	/**
-	 * 
-	 * @author Todor Balabanov
-	 * 
-	 * @email tdb@tbsoft.eu
-	 * 
-	 * @date 09 Aug 2012
-	 */
-	public Game() {
-		for (int i = 0; i < players.length; i++) {
-			players[i] = new Player();
-		}
-	}
-}
-
-/**
- * 
- * @author Todor Balabanov
- * 
- * @email tdb@tbsoft.eu
- * 
- * @date 09 Aug 2012
- */
-class HandStrength {
-	long fifthKicker;
-	long fourthKicker;
-	long thirdKicker;
-	long secondKicker;
-	long firstKicker;
-	long onePair;
-	long twoPair;
-	long threeOfKind;
-	long straight;
-	long flush;
-	long fullHouse;
-	long fourOfKind;
-	long straightFlush;
-
-	/**
-	 * 
-	 * @return
-	 * 
-	 * @author Todor Balabanov
-	 * 
-	 * @email tdb@tbsoft.eu
-	 * 
-	 * @date 09 Aug 2012
-	 */
-	long getValue() {
-		long result = 0L;
-
-		int offset = 0;
-
-		/*
-		 * long fifthKicker:(4);
-		 */
-		result |= (fifthKicker << offset);
-		offset += 4;
-
-		/*
-		 * long fourthKicker:(4);
-		 */
-		result |= (fourthKicker << offset);
-		offset += 4;
-
-		/*
-		 * long thirdKicker:(4);
-		 */
-		result |= (thirdKicker << offset);
-		offset += 4;
-
-		/*
-		 * long secondKicker:(4);
-		 */
-		result |= (secondKicker << offset);
-		offset += 4;
-
-		/*
-		 * long firstKicker:(4);
-		 */
-		result |= (firstKicker << offset);
-		offset += 4;
-
-		/*
-		 * long onePair:(4);
-		 */
-		result |= (onePair << offset);
-		offset += 4;
-
-		/*
-		 * long twoPair:(4);
-		 */
-		result |= (twoPair << offset);
-		offset += 4;
-
-		/*
-		 * long threeOfKind:(4);
-		 */
-		result |= (threeOfKind << offset);
-		offset += 4;
-
-		/*
-		 * long straight:(4);
-		 */
-		result |= (straight << offset);
-		offset += 4;
-
-		/*
-		 * long flush:(1);
-		 */
-		result |= (flush << offset);
-		offset += 1;
-
-		/*
-		 * long fullHouse:(1);
-		 */
-		result |= (fullHouse << offset);
-		offset += 1;
-
-		/*
-		 * long fourOfKind:(1);
-		 */
-		result |= (fourOfKind << offset);
-		offset += 1;
-
-		/*
-		 * long straightFlush:(1);
-		 */
-		result |= (straightFlush << offset);
-		offset += 1;
-
-		return (result);
-	}
-}
-
-/**
- * 
- * @author Todor Balabanov
- * 
- * @email tdb@tbsoft.eu
- * 
- * @date 09 Aug 2012
- */
 public class MonteCarlo {
 
 	/**
