+++ src/eu/veldsoft/vitosha/poker/odds/model/Pot.java
@@ -1,0 +1,14 @@
+package eu.veldsoft.vitosha.poker.odds.model;
+
+/**
+ * 
+ * @author Todor Balabanov
+ * 
+ * @email tdb@tbsoft.eu
+ * 
+ * @date 09 Aug 2012
+ */
+class Pot {
+	long[] bets = new long[Constants.MAX_NUMBER_OF_PLAYERS];
+	int numberOfSplits;
+}
